Updated project structure and container component
