+++ src/pages.js
@@ -1,0 +1,4 @@
+export { default as Dashboard } from 'pages/Dashboard';
+export { default as Error } from 'pages/Error';
+export { default as Login } from 'pages/Login';
+export { default as Selector } from 'pages/Selector';
