+++ src/pages/Error.jsx
@@ -1,0 +1,7 @@
+function Error() {
+  return (
+    <div>Error</div>
+  )
+}
+
+export default Error
