+++ src/pages/Login.jsx
@@ -1,0 +1,5 @@
+function Login() {
+  return <div>Login</div>;
+}
+
+export default Login;
