+++ src/pages/Selector.jsx
@@ -1,0 +1,5 @@
+function Selector() {
+  return <div>Selector</div>;
+}
+
+export default Selector;
